+++ src/test/indexTest.ts
@@ -1,6 +1,5 @@
 import chai = require('chai');
 const expect = chai.expect;
-const assert = chai.assert;
 
 import childProcess = require('child_process');
 const exec = childProcess.exec;
@@ -56,6 +55,20 @@ describe('index', () => {
                 expect(stdout).to.contain('0 crit');
                 done();
             });
+        }).timeout(5000);
+
+        it('validate importValue flag', (done) => {
+            exec('node lib/index.js validate testData/valid/yaml/valid_import_value_type.yaml --import-values ImportedValueOutputPort="20"', function(error, stdout, stderr) {
+                expect(stdout).to.contain('0 crit');
+                done();
+            })
+        }).timeout(5000);
+
+        it('validate numeric importValue flag', (done) => {
+            exec('node lib/index.js validate testData/valid/yaml/valid_import_value_type.yaml --import-values ImportedValueOutputPort=20', function(error, stdout, stderr) {
+                expect(stdout).to.contain('0 crit');
+                done();
+            })
         }).timeout(5000);
 
         it('validate pseudo flag', (done) => {